finish entities relatiosn

+++ src/main/java/fr/slghive/heartlink/entities/TypeEntity.java
@@ -15,8 +15,9 @@ public class TypeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
 
     @Column(nullable = false)
     private String name;
+
 }
